FRONT : Renamed Auth folder and moved it into components folder, further file organization
