+++ scratch-and-map-front-end/src/components/AuthContainer/Auth.js
