+++ scratch-and-map-front-end/src/components/AuthContainer/Register.js
