+++ scratch-and-map-front-end/src/components/NavContainer/NavBar.js
@@ -1,6 +1,5 @@
 import React from "react";
-import Auth from "../../Auth/Auth";
-
+import Auth from "../AuthContainer/Auth";
 import { Menu, Sidebar, Button, Icon, Dropdown } from "semantic-ui-react";
 import SidebarDrop from "./SidebarDrop";
 import SelectLanguage from "./SelectLanguage";
